requirements.txt
asgiref==3.7.2
attrs==23.2.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.8
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
phonenumbers==8.13.27
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyrsistent==0.20.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
